pyproject.toml
[project]
name = "tech-radar-scraper"
version = "0.1.0"
description = "自动化技术雷达 - 每日 .NET 博客与 GitHub 热门项目聚合"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0.11",   # 解析 RSS/Atom 订阅源
    "requests>=2.32.0",     # HTTP 请求（GitHub API）
    "python-dotenv>=1.0.0", # 从 .env 文件加载环境变量
]

[dependency-groups]
dev = []
